Import 64-bit fixes from dcload-ip
This was fixed by BlueCrab in 2013 for dcload-ip
and just never made it into dcload-serial. The issue
is that the original dcload was set up to just sent
the raw host-side `DIR *` over to the target. This
worked fine when those pointers were 32-bit, but as
hosts moved to 64-bit it simply broke.

Now instead, we keep a cache of those pointers and
assign an fd before sending over to the target.
This is maintaining the offset like dcload-ip, but
a PR will be forthcoming to update fs_dcload to
undo the hack on the dc side.